pyproject.toml
[project]
name = "money-plant-aptos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["aptos-sdk>=0.11.0",    "python-dotenv>=1.0.0,<2"]
